Minor Updates And BugFixes
- Fixed on `Str::slug` method to properly filter slugs.
- Fixed on `Pixl::crop` method to crop images from center correctly.
- Fixed on `Post->author` magic method to set/get a new author when the author is changed on existing posts, before the post is resaved.
- Fixed `Media::delete` method to remove thumbnail_id from posts when the media item is deleted.
- The CMS no longer routes all attachment sizes. Only the primary attachement is routed.
- PHPDocBlock update on `Coupons::used` method.
- Fixed the writer in the admin panel to set the correct author on existing posts.
- Updated logic for listing leads in the admin panel for faster parsing.
- `MediaLibrary::delete` checks if the attachement exists before deleting it in the admin panel
- Added `visited-checkout` to the lead filters in the admin panel

+++ kanso/cms/admin/models/MediaLibrary.php
@@ -173,7 +173,10 @@ class MediaLibrary extends BaseModel
         {
             $attachment = $this->MediaManager->byId($id);
 
-            $attachment->delete();
+            if ($attachment)
+            {
+                $attachment->delete();
+            }
         }
 
         return 'valid';
